Code completely broken. Rearringaing some interfaces, removing one use if, they do not add any functionality.
