+++ src/enjug/erijan/games/yatzy/RollControl.java
@@ -4,6 +4,6 @@ package enjug.erijan.games.yatzy;
  * Created by Jan Eriksson on 16/11/15.
  */
 public interface RollControl {
-  void rollActiveDice();
+  void rollActiveDice(String playerName);
   int rollsLeft();
 }
