+++ src/enjug/erijan/games/yatzy/RollControlYatzy.java
@@ -14,15 +14,41 @@ public class RollControlYatzy implements RollControl {
   private DiceHandler diceHandler;
   private List<Player> players;
   private int reRolls;
+  private int rollsDone;
 //  Map<String,Integer> rolls
 
-  RollControlYatzy(List<Player> players, DiceHandler diceHandler) {
-    this.players = players;
+  RollControlYatzy(DiceHandler diceHandle) {
     this.diceHandler = diceHandler;
   }
 
   @Override
-  public void rollActiveDice() {
+  public void rollActiveDice(String playerName) {
+    String message;
+    if (rollsDone == 0) {
+      diceHandler.setAllDiceActive();
+      rollsDone++;
+      diceHandler.rollActiveDice();
+      setTempScore();
+      message = playerName + " rolled some dice.";
+
+    } else if (rollsDone < noOfReRolls) {
+      rollsDone++;
+      diceHandler.rollActiveDice();
+      setTempScore();
+      message = playerName + " rolled some dice.";
+
+    } else if (rollsDone == noOfReRolls) {
+      rollsDone++;
+      diceHandler.rollActiveDice();
+      setTempScore();
+      message = playerName + " rolled some dice.";
+      diceHandler.deActivateAllDice();
+
+    } else {
+      message = "No more rolls alowed.";
+    }
+    //return message;
+    stateInfo.setStateMessage(message);
     diceHandler.rollActiveDice();
   }
