+++ src/enjug/erijan/games/yatzy/VariantFactory.java
@@ -7,7 +7,7 @@ import enjug.erijan.games.yatzy.view.YatzyGui;
  * Created by Jan Eriksson on 18/11/15.
  */
 public interface VariantFactory {
-  ScoreColumn getScoreModel(Player player);
+  ScoreReader getScoreModel(Player player);
   DiceHandler getDice();
   YatzyGui getGui(GameControl yatzyAgent, DiceHandler diceHandler, StateInfo stateInfo);
 }
